added main and updated cargo toml
